Watch for NPE when deleting fields.
Added a check for NPE when fields are being deleted.  Still,
the Manage Credits section should be re-examined to make sure
that fields are remaining in-sync, but the error should no
longer manifest.

Ready to Deploy #724

Reviewed-on: http://review.source.gogoego.com/3599
Tested-by: Carl Scott <carl.scott@solertium.com>

+++ org.iucn.sis.server/src/org/iucn/sis/server/restlets/assessments/AssessmentChangesRestlet.java
@@ -289,17 +289,19 @@ public class AssessmentChangesRestlet extends BaseServiceRestlet {
 				continue;
 			}
 			
-			saver.addChange(saver.createDeleteChange(toDelete));
-			
-			try {
-				FieldDAO.deleteAndDissociate(toDelete, session);
-			} catch (PersistentException e) {
-				Debug.println(e);
-				throw new ResourceException(Status.SERVER_ERROR_INTERNAL, e);
+			if (toDelete != null) {
+				saver.addChange(saver.createDeleteChange(toDelete));
+				
+				try {
+					FieldDAO.deleteAndDissociate(toDelete, session);
+				} catch (PersistentException e) {
+					Debug.println(e);
+					throw new ResourceException(Status.SERVER_ERROR_INTERNAL, e);
+				}
+				assessment.getField().remove(toDelete);
+				
+				out.append(XMLWritingUtils.writeTag("li", "Deleted empty field " + field.getName()));
 			}
-			assessment.getField().remove(toDelete);
-			
-			out.append(XMLWritingUtils.writeTag("li", "Deleted empty field " + field.getName()));
 		}
 		
 		try {
